# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Bancaire",
        "description": "API REST pour la gestion des comptes bancaires",
        "contact": {
            "email": "contact@ramatoulaye.gueye.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-ramatoulaye-gueye-i671.onrender.com/api/v1",
            "description": "Serveur de production (Render)"
        },
        {
            "url": "https://api-ramatoulaye-gueye-i671.onrender.com/api/v1",
            "description": "Serveur de production alternatif"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Serveur de développement local"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement alternatif"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifie un utilisateur (Admin ou Client) et retourne les tokens d'accès",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@banque.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe",
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nom": {
                                                            "type": "string",
                                                            "example": "Dupont"
                                                        },
                                                        "prenom": {
                                                            "type": "string",
                                                            "example": "Jean"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin@banque.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "enum": [
                                                                "admin",
                                                                "client"
                                                            ],
                                                            "example": "admin"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Identifiants invalides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/comptes/{compteId}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer un compte spécifique",
                "description": "Récupère les détails d'un compte bancaire spécifique selon les permissions de l'utilisateur",
                "operationId": "getCompte",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte bancaire ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "b6500996-a594-495b-ab68-c3727094f52d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Données récupérées avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "numeroCompte": {
                                                    "type": "string",
                                                    "example": "CPT1761572199795"
                                                },
                                                "titulaire": {
                                                    "type": "string",
                                                    "example": "Amadou Diallo"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "epargne",
                                                        "cheque"
                                                    ]
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1250000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                },
                                                "dateCreation": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "enum": [
                                                        "actif",
                                                        "bloque",
                                                        "ferme"
                                                    ]
                                                },
                                                "motifBlocage": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "metadata": {
                                                    "properties": {
                                                        "derniereModification": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "version": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "COMPTE_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Le compte avec l'ID spécifié n'existe pas"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "compteId": {
                                                            "type": "string",
                                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètres invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Token manquant ou invalide"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Données de validation invalides"
                    },
                    "429": {
                        "description": "Trop de requêtes - Rate limiting"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte bancaire",
                "description": "Supprime un compte bancaire avec un soft delete. Le compte passe au statut 'ferme' et n'est plus visible dans les listes normales. Seuls les comptes actifs peuvent être supprimés.",
                "operationId": "deleteCompte",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte bancaire (doit être un compte actif)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "b6500996-a594-495b-ab68-c3727094f52d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "numeroCompte": {
                                                    "type": "string",
                                                    "example": "CPT1761572199795"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "ferme"
                                                },
                                                "dateFermeture": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-10-19T11:15:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "COMPTE_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Le compte avec l'ID spécifié n'existe pas"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "compteId": {
                                                            "type": "string",
                                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Token manquant ou invalide"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}